Add timer update. Add new step sensor. Auto select step sensor by API and supported features. Work detector in sleep mode. Integrating with model. Save data in database. Load data from database (Problem class cast exception). Need more test.

+++ app/src/main/java/com/intrafab/medicus/pedometer/OnEventUpdateCalories.java
@@ -5,9 +5,9 @@ package com.intrafab.medicus.pedometer;
  * Copyright (c) 2015 Artemiy Terekhov. All rights reserved.
  */
 public class OnEventUpdateCalories {
-    public final double calories;
+    public final float calories;
 
-    public OnEventUpdateCalories(double calories) {
+    public OnEventUpdateCalories(float calories) {
         this.calories = calories;
     }
 }
